Bugfix - Reverse sort uris + release 1.3.1
URIs used to be added in alphabetical order

    /user/:userId
    /user/:userId/send-email
    /user/:userId/project/:projectId
    /user/:userId/status

Unfortunately, that meant the first route would get the request for at
least one framework even if the request was for the second or third
route. Now routes are sorted by the number of elements in the path
(descending) and then alphabetically.

    /user/:userId/project/:projectId
    /user/:userId/send-email
    /user/:userId/status
    /user/:userId